implement support for remarkable 1.0
I've noticed from the pipes-and-paper project that the only difference
between ReMarkable 1.0 and 2.0 is the event file. I've implemented this
in the rust code here so I can use it on my remarkable.

I'm pretty sure there is a better way to handle the different hardware
versions but I'm no rust expert...